+++ activerecord/lib/active_record/relation/query_methods.rb
@@ -1073,13 +1073,14 @@ module ActiveRecord
     end
 
     # Checks to make sure that the arguments are not blank. Note that if some
-    # blank-like object were initially passed into the query method, then this
-    # method will not raise an error.
+    # blank-like not nil object were initially passed into the query method,
+    # then this method will not raise an error.
     #
     # Example:
     #
-    #    Post.references()   # => raises an error
-    #    Post.references([]) # => does not raise an error
+    #    Post.references()    # => raises an error
+    #    Post.references(nil) # => raises an error
+    #    Post.references([])  # => does not raise an error
     #
     # This particular method should be called with a method_name and the args
     # passed into that method as an input. For example:
@@ -1089,7 +1090,7 @@ module ActiveRecord
     #   ...
     # end
     def check_if_method_has_arguments!(method_name, args)
-      if args.blank?
+      if args.blank? || (method_name != :reorder && args.compact.blank?)
         raise ArgumentError, "The method .#{method_name}() must contain arguments."
       end
     end